+++ gfxutil.cpp
@@ -16,7 +16,7 @@
 
 #include <math.h>
 
-int getvsyncrate (int hz)
+double getvsyncrate (double hz)
 {
 	if (hz > 85)
 		return hz / 2;
