+++ od-win32/sounddep/sound.cpp
@@ -150,12 +150,12 @@ static int isvsync (void)
 
 float scaled_sample_evtime_orig;
 extern float sampler_evtime;
-void update_sound (int freq, int longframe, int linetoggle)
+void update_sound (double freq, int longframe, int linetoggle)
 
 {
-	static int lastfreq;
-	float lines = 0;
-	float hpos;
+	static double lastfreq;
+	double lines = 0;
+	double hpos;
 
 	if (freq < 0)
 		freq = lastfreq;
@@ -175,7 +175,7 @@ void update_sound (int freq, int longframe, int linetoggle)
 		hpos = maxhpos_short;
 	}
 	lines += maxvpos_nom;
-	scaled_sample_evtime_orig = hpos * lines * freq * CYCLE_UNIT / (float)sdp->obtainedfreq;
+	scaled_sample_evtime_orig = hpos * lines * freq * CYCLE_UNIT / (double)sdp->obtainedfreq;
 	scaled_sample_evtime = scaled_sample_evtime_orig;
 	sampler_evtime = hpos * lines * freq * CYCLE_UNIT;
 #if 0
